testy vagranta do konca
